requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
black==23.3.0
blessed==1.20.0
cached-property==1.5.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dj-stripe==2.7.3
Django==4.2.2
django-debug-toolbar==4.1.0
django-enum==1.2.1
django-enumfields==2.1.1
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
enum-compat==0.0.3
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.17.3
magic-filter==1.0.9
mixins==0.1.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
redis==3.5.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==4.2.0
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
utils==1.0.1
wcwidth==0.2.6
yarl==1.9.2
